==> 2_PBN_Geocoding.ipynb <==
import pandas as pd

df = pd.read_csv('/content/drive/MyDrive/Capstone Project: PBN/Data/Original/company_data.csv')

df.head()

def post_to_region(n):
  if n>=1011 and n<=1806:
    return('Budapest')
  elif n>=2000 and n<=2769:
    return('Pest')
  elif (n>=2400 and n<=2490) or (n>=8000 and n<=8157):
    return('Fejér County')
  elif (n>=2500 and n<=2545) or (n>=2800 and n<=2949):
    return('Komárom-Esztergom County')
  elif (n>=2640 and n<=2699) or (n>=3041 and n<=3253):
    return('Nógrád County')
  elif (n>=3000 and n<=3036) or (n>=3200 and n<=3399):
    return('Heves County')
  elif n>=3400 and n<=3999:
    return('Borsod-Abaúj-Zemplén County')
  elif (n>=4000 and n<=4288):
    return('Hajdú-Bihar County')
  elif (n>=4300 and n<=4977):
    return('Szabolcs-Szatmár-Bereg County')
  elif (n>=5000 and n<=5476):
    return('Jász-Nagykun-Szolnok County')
  elif (n>=5500 and n<=5948):
    return('Békés County')
  elif (n>=6000 and n<=6528):
    return('Bács-Kiskun County')
  elif (n>=6600 and n<=6932):
    return('Csongrád-Csanád County')
  elif (n>=7020 and n<=7228):
    return('Tolna County')
  elif (n>=7300 and n<=7396) or (n>=7600 and n<=7985):
    return('Baranya County')
  elif (n>=7400 and n<=7589) or (n>=8600 and n<=8739):
    return('Somogy County')
  elif (n>=8161 and n<=8598):
    return('Veszprém County')
  elif (n>=8353 and n<=8395) or (n>=8741 and n<=8999):
    return('Zala County')
  elif (n>=9001 and n<=9495):
    return('Győr-Moson-Sopron County')
  elif (n>=9500 and n<=9985):
    return('Vas County')
  else:
    return('NaN')




df['Postal_Code'] = pd.to_numeric(df['Postal_Code'], errors='coerce')

df['region'] = [post_to_region(n) for n in df['Postal_Code']]

df['region'].value_counts()

def region_to_nuts2(n):
  if n in ['Budapest', 'Pest']:
    return(n)
  elif n in ['Fejér County', 'Komárom-Esztergom County', 'Veszprém County']:
    return('Central Transdanubia')
  elif n in ['Győr-Moson-Sopron County', 'Vas County', 'Zala County']:
    return('Western Transdanubia')
  elif n in['Baranya County','Somogy County','Tolna County']:
    return('Southern Transdanubia')
  elif n in ['Borsod-Abaúj-Zemplén County','Heves County','Nógrád County']:
    return('Northern Hungary')
  elif n in ['Hajdú-Bihar County', 'Jász-Nagykun-Szolnok County', 'Szabolcs-Szatmár-Bereg County']:
    return('Northern Great Plain')
  elif n in ['Bács-Kiskun County', 'Békés County', 'Csongrád-Csanád County']:
    return('Southern Great Plain')
  else:
    return('NaN')

df['NUTS2'] = [region_to_nuts2(n) for n in df['region']]

df['NUTS2'].value_counts()

df.head()

df.to_csv('/content/drive/MyDrive/Capstone Project: PBN/Data/company_data_with_region.csv')

df.groupby('Industry')['Tax_number'].count()

merged = pd.read_csv('/content/drive/MyDrive/Capstone Project: PBN/Data/merged_company_financial.csv')

merged.groupby('sector')['Tax_number'].count()
