==> 3_PBN_Financial_Ratios.ipynb <==
import pandas as pd

financials_clean = pd.read_csv("/content/drive/MyDrive/Capstone Project: PBN/Data/financials_clean.csv")

financials_clean.columns

financials_clean["ROA"]= financials_clean['Profit After Tax IS']/financials_clean['Total Assets'] 

financials_clean["ROA"]

financials_clean["Inventory_turnover"]=financials_clean['Cost of Goods Sold']/financials_clean['Inventory']

financials_clean["Inventory_turnover"]

financials_clean['sum']=financials_clean[['Material Cost','Cost of Services','Cost of Goods Sold','Salary Expenses','Depreciation','Other Expenses','Financial Operations Cost']].fillna(0).sum(axis=1)

sum_ind = financials_clean.columns.get_loc('sum')
COGS_ind = financials_clean.columns.get_loc('Cost of Goods Sold')

zero_sum_ind = financials_clean[financials_clean['sum']==0].index

non_zero_ind = list(set(financials_clean.index)^set(zero_sum_ind))

financials_clean["Manufacturing_Costs_to_Total_Expenses"] = 0

financials_clean.iloc[non_zero_ind, financials_clean.columns.get_loc("Manufacturing_Costs_to_Total_Expenses")] = financials_clean.iloc[non_zero_ind, COGS_ind]/financials_clean.iloc[non_zero_ind,sum_ind]

financials_clean["Manufacturing_Costs_to_Total_Expenses"].value_counts()

financials_clean.drop(columns = 'sum', inplace=True)

financials_clean.drop(columns = 'Unnamed: 0', inplace=True)

financials_clean.head()

financials_clean.rename({'Assets':'Fixed Assets'}, axis = 1, inplace=True)

financials_clean.columns

list_of_var = ['Total Net Sales', 
               'Cost of Goods Sold', 
               'Gross Income', 
               'Financial Operations Income', 
               'Regular Operations Income', 
               'Extraordinary Income', 
               'Profit After Tax IS', 
               'Profit or Loss of The Year IS', 
               'Total Assets',
               "Shareholder's Equity",
               'Long-Term Liabilities',
               'Current Liabilities',
               'Total Liabilities',
               'ROA',
               'Inventory_turnover',
               'Manufacturing_Costs_to_Total_Expenses'
               ]

financials_clean.to_csv('/content/drive/MyDrive/Capstone Project: PBN/Data/financials_index_added.csv')
