==> 4_PBN_Grouping.ipynb <==
import pandas as pd
import numpy as np
import os

company_indgroup = pd.read_csv("company_data_with_region.csv")
company_indgroup.info()

# create a list of our conditions
conditions = [
    (company_indgroup['Industry'] == 'Food Production'), 
    (company_indgroup['Industry'] == 'Manufacture of beverages'),
    (company_indgroup['Industry'] == 'Manufacture of textiles'), 
    (company_indgroup['Industry'] == 'Manufacture of wearing apparel'), 
    (company_indgroup['Industry'] == 'Manufacture of leather and related products, footwear'),
    (company_indgroup['Industry'] == 'Manufacture of wood and of products of wood and cork, except furniture;'), 
    (company_indgroup['Industry'] == 'Manufacture of paper and paper products'), 
    (company_indgroup['Industry'] == 'Printing and other reproduction activities'), 
    (company_indgroup['Industry'] == 'Furniture production'),
    (company_indgroup['Industry'] == 'Manufacture of rubber and plastic products'), 
    (company_indgroup['Industry'] == 'Manufacture of non - metallic mineral products'), 
    (company_indgroup['Industry'] == 'Manufacture of basic metals'),
    (company_indgroup['Industry'] == 'Manufacture of fabricated metal products'),
    (company_indgroup['Industry'] == 'Manufacture of computer, electronic and optical products'), 
    (company_indgroup['Industry'] == 'Manufacture of electrical equipment'),
    (company_indgroup['Industry'] == 'Manufacture of machinery and equipment nec'), 
    (company_indgroup['Industry'] == 'Manufacture of road vehicles'), 
    (company_indgroup['Industry'] == 'Manufacture of other transport equipment'),
    (company_indgroup['Industry'] == 'Other manufacturing')]

# create a list of the values we want to assign for each condition
values = ['agri', 'agri', 
          'text_leather', 'text_leather', 'text_leather', 
          'wood_paper_furni', 'wood_paper_furni', 'wood_paper_furni', 'wood_paper_furni', 
          'chemicals', 'chemicals',
          'basic_metals', 'basic_metals',
          'comp_electr', 'comp_electr',
          'machinery','machinery','machinery', 
          'other']

# create a new column and use np.select to assign values to it using our lists as arguments
company_indgroup['sector'] = np.select(conditions, values, 'non-mnfct')

# check results
company_indgroup.sector.unique()

company_indgroup.to_csv('check_company.csv')

# IMPORTANT ASSUMPTION 1: every company's HQ is recorded in the company data set. No firms have only site branches. 
# In other words, the line below would not remove companies for the lack of HQ branch. (Only 3 tax_numbers don't have HQ)
# IMPORTANT ASSUMPTION 2: In the financial data set, every company has only 1 record of financial values each year. 
# In other words, when matching up company info and financials, it would be 1:1 relationship each year. (true)

# select columns for new company data frame
company = company_indgroup[['Tax_number', 'NUTS2', 'sector', 'HQ_Site_Branch']]
# remove duplicates
company.drop_duplicates(keep=False,inplace=True)
# remove non-manufacturers
company.drop(company[company['sector'] == 'non-mnfct'].index, inplace = True)
# remove non-HQ
company.drop(company[company['HQ_Site_Branch'] != 'HQ'].index, inplace = True)
# check results
company

# read in financial data
financial_ori = pd.read_csv("financials_index_added.csv")
financial_ori

# make ID column name the same
financial_ori.rename(columns = {'Tax Number':'Tax_number'}, inplace = True)
# select columns for new finance data frame
financial = financial_ori[['Tax_number', 'Report Start Date', 'Report End Date',
                'Total Net Sales', 
               'Cost of Goods Sold', 
               'Gross Income', 
               'Financial Operations Income', 
               'Regular Operations Income', 
               'Extraordinary Income', 
               'Profit After Tax IS', 
               'Profit or Loss of The Year IS', 
               'Total Assets',
               "Shareholder's Equity",
               'Long-Term Liabilities',
               'Current Liabilities',
               'Total Liabilities',
               'ROA',
               'Inventory_turnover',
               'Manufacturing_Costs_to_Total_Expenses']]
# remove duplicates
financial.drop_duplicates(keep=False,inplace=True)
# check results
financial

financial.isnull().sum().sort_values(ascending=True)

# change ID column data type
financial['Tax_number'] = financial['Tax_number'].astype('category')
company['Tax_number'] = company['Tax_number'].astype('category')

# merge the two data frames
df = pd.merge(company, financial, how = 'inner', on = 'Tax_number')
# remove duplicates
df.drop_duplicates(keep=False,inplace=True)
# check results
df

df.to_csv('NEW_merged_company_financial.csv')

# Check if Profit After Tax is a good indicator (not many nans)
df.isnull().sum().sort_values(ascending=True)

# Check if Tableau's calculation for averages skips nans
check = df.query('`Report Start Date`>="2020-01-01" and `Report Start Date`<="2020-12-31" and sector == "chemicals"')

check

pd.set_option('display.float_format', lambda x: '%.3f' % x)
check.drop(columns=['Tax_number','NUTS2','sector','Report Start Date','Report End Date']).mean()

check.isnull().sum()
