==> 5_PBN_Resilience.ipynb <==
import pandas as pd
import numpy as np
import datetime as dt

df = pd.read_csv("NEW_merged_company_financial.csv")
df.head()

# Drop columns
df = df.drop(columns=["Unnamed: 0", 'Report End Date', 'HQ_Site_Branch'])

# convert tax number column type
df['Tax_number'] = pd.Categorical(df.Tax_number)

# convert date column type
df['Report Start Date'] = pd.to_datetime(df['Report Start Date'])

# unitfy time unit to year
df['Report Start Date'] = df['Report Start Date'].dt.strftime('%Y')
df.head()

# unify monthly/quaterly profit values into annual ones
annual_profit = pd.DataFrame(df.groupby(['Tax_number', 'Report Start Date'])['Profit After Tax IS'].sum())

# rename column to annual_profit
annual_profit.rename(columns = {'Profit After Tax IS': 'annual_profit'}, inplace=True)

# check the result
annual_profit.head()
# Note: groupby() adds non-existing years with 0.0

# Normalize annual_profit
# 1. replace 0.0 with nans. Otherwise, avergaes will be dragged down. 
annual_profit['annual_profit'].replace(0.0, np.NaN)

# 2. calculate avg annual_profit 
profit_mean = pd.DataFrame(annual_profit.groupby('Tax_number')['annual_profit'].mean())
profit_mean.rename(columns = {'annual_profit': 'firm_avg_annual_profit'}, inplace=True)

# 3. merge profit_mean with df
df = df.merge(profit_mean, how="left", on = 'Tax_number')

# 4. merge annual_profit with df
df = df.merge(annual_profit, how="left", on = ['Tax_number', 'Report Start Date'])

# 5. normalized = annual_profit / firm_avg_annual_profit
df['norm_annual_profit'] = df['annual_profit'] / df['firm_avg_annual_profit']

# check the result
df.head()

# Get annual profit changes
# 1. group by for avg, to avoid duplicated norm_annual_profit
df1 = pd.DataFrame(df.groupby(['Tax_number', 'Report Start Date'])['norm_annual_profit'].mean())

# 2. calculate difference between rows by group
df1['change_annual_profit'] = df1['norm_annual_profit'].diff()

# 3. merge df1 with df
df = df.merge(df1, how="left", on = ['Tax_number', 'Report Start Date'])

# 4. check the result
df.head()

# Calculate industrial avg
# 1. group by
industrial_avg = pd.DataFrame(df.groupby(['sector','Report Start Date'])['norm_annual_profit_x'].mean())

# 2. rename column to industrial avg
industrial_avg.rename(columns = {'norm_annual_profit_x': 'industrial_avg'}, inplace=True)

# 3. get industrial avg changes
industrial_avg['change_industrial_avg'] = industrial_avg['industrial_avg'].diff()

# 4. merge industrial_avg with df
df = df.merge(industrial_avg, how="left", on = ['sector','Report Start Date'])
# ? How to do differencing within groups

# 5. check the result
df.head()

df.isnull().sum().sort_values(ascending=True)
# ? why there are so many Nans in ind avg

# create a list of conditions
conditions = [
    (df['change_industrial_avg'] < 0) & (df['change_annual_profit'] > 0),
    (df['change_industrial_avg'] == 0) & (df['change_annual_profit'] > 0),
    (df['change_industrial_avg'] < 0) & (df['change_annual_profit'] == 0),
    (df['change_industrial_avg'] > 0) & (df['change_annual_profit'] > 0),
    (df['change_industrial_avg'] == 0) & (df['change_annual_profit'] == 0),
    (df['change_industrial_avg'] < 0) & (df['change_annual_profit'] < 0),
    (df['change_industrial_avg'] > 0) & (df['change_annual_profit'] == 0),
    (df['change_industrial_avg'] == 0) & (df['change_annual_profit'] < 0),
    (df['change_industrial_avg'] > 0) & (df['change_annual_profit'] < 0)
    ]
# create a list of the values to assign for each condition
values = [5, 
          4, 4, 
          3, 3, 3, 
          2, 2,
          1]
# create a new column and use np.select to assign values to it using our lists as arguments
df['reactive_resilience'] = np.select(conditions, values)

# 0 represents Nan, fill 0 with nan
df['reactive_resilience'].replace(0, np.NaN)

# check results
df.head()

# Get a simplified dataframe
finaldf = df[['sector', 'NUTS2', 'Tax_number', 'Report Start Date', 'reactive_resilience', 
               'Total Net Sales', 
               'Cost of Goods Sold', 
               'Gross Income', 
               'Financial Operations Income', 
               'Regular Operations Income', 
               'Extraordinary Income', 
               'Profit After Tax IS', 
               'Profit or Loss of The Year IS', 
               'Total Assets',
               "Shareholder's Equity",
               'Long-Term Liabilities',
               'Current Liabilities',
               'Total Liabilities',
               'ROA',
               'Inventory_turnover',
               'Manufacturing_Costs_to_Total_Expenses']]
finaldf.head()

finaldf.to_csv('df_for_modelling.csv')

# Get firms worth investment
# 1. calculate average resilience scores
avg_resilience = pd.DataFrame(finaldf.groupby('Tax_number')['reactive_resilience'].mean())
avg_resilience.rename(columns = {'reactive_resilience': 'avg_resilience'}, inplace=True)

# 2. keep existing firms
# 2.1 get a list of firms that still exist
firm_list = finaldf[finaldf['Report Start Date'] == '2020']
firm_list = firm_list[['sector', 'NUTS2', 'Tax_number']]

# 2.2 exclude firms not on the list by inner joining avg_resilience with firm_list
final_firm_list = pd.merge(firm_list, avg_resilience, how = 'inner', on = 'Tax_number')

# check the result
final_firm_list.head()

# rank firms within industries
final_firm_list["rank_per_ind"] = final_firm_list.groupby("sector")["avg_resilience"].rank("dense", ascending=False)

# select NO.1 in each industry
top_firms_per_ind = final_firm_list[final_firm_list['rank_per_ind'] == 1]
top_firms_per_ind.drop_duplicates(subset=['Tax_number'], inplace = True)
top_firms_per_ind

# rank firms within regions
final_firm_list["rank_per_reg"] = final_firm_list.groupby("NUTS2")["avg_resilience"].rank("dense", ascending=False)

# select NO.1 in each industry
top_firms_per_reg = final_firm_list[final_firm_list['rank_per_reg'] == 1]
top_firms_per_reg.drop_duplicates(subset=['Tax_number'], inplace = True)
top_firms_per_reg
